extended universal intrinsics: added 64-bit integers, improved shift operators, renamed "pack" operations; changed semantics of extended integer multiplication, added v_dotprod. also, split intrinsics into separate files. C implementation is compiled only when none of SSE2 or NEON is available.
